refactor: extend server.js for components
